Use updated server API (#14)
Instead of returning the full url to the binary directory now returns
`bin_dir` only. This is possible as `asu_url` is already defined
locally. This approach allows devices not using https to download
content unencrypted, meaning the server response does not delegate the
protocol.

In addition a boolean value is returned stating if a buildlog was
created, which is shown in case of failures.

Signed-off-by: Paul Spooren <mail@aparcar.org>