fix autocomplete + buildate, move exampels
autocomplete never worked because of wrong files naming. The package
index file is called `index.json` and now available on the ASU server.

The build date never worked for images created by the official OpenWrt
build bots, fixed by reading the `last-modified` header.

Examples files are now in their own folders to follow the structure of
ASU, which offers not only profiles but also package information in the
version folders.

Signed-off-by: Paul Spooren <mail@aparcar.org>